+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-- Changed auth strategy to use `find_or_create_token` endpoint, deprecated `auth.refresh_tokens()` ([#484](https://github.com/nsidc/earthaccess/issues/484)) ([**@fwfichtner**](https://github.com/fwfichtner)
+- Automatically refresh EDL token and deprecate the `Auth.refresh_tokens` method with no replacement, as there is no longer a need to explicitly refresh ([#484](https://github.com/nsidc/earthaccess/issues/484)) ([**@fwfichtner**](https://github.com/fwfichtner))
 - Deprecate `earthaccess.get_s3fs_session` and `Store.get_s3fs_session`. Use
   `earthaccess.get_s3_filesystem` and `Store.get_s3_filesystem`, respectively,
   instead ([#766](https://github.com/nsidc/earthaccess/issues/766))([**@Sherwin-14**](https://github.com/Sherwin-14))